added static forRoot method
usually you want a default for your CgBusyDefaults for the whole application.
Since it got "provided" every time you import the module
(which you must if you want to use the directive in a
lazy loaded module or a submodule) the defaults got
overwritten everytime.

Now it is provided in forRoot which has to be called in your app.module.

This is a breaking change!